One file is enough to give a little introduction about lists
This code is for learning only i took inspiration from an italian channel on yt. Here's the video: https://youtu.be/Qhh5TcDudTw
The code is all commented in Italian, sorry for the not bilingualism but i preferred to write the comment in my mother language. 
Maybe in the future i will upload it all commented in english.